SDHR-Guard/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
python-multipart==0.0.6
aiofiles==23.2.1
pydantic>=2.0.0
psutil>=5.9.0
pydantic-settings>=2.1.0
typing-extensions>=4.8.0 